Image_processing/requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.1.2
flaskwebgui==0.3.5
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
Werkzeug==2.1.2
whichcraft==0.6.1
zipp==3.8.0
